Authorization-aware control-plane components
The control-plane components relied on a `--single-namespace`, passed
from `linkerd install` into each individual component, to determine
which namespaces they were authorized to access, and whether to support
ServiceProfiles. This command-line flag was redundant given the
authorization rules encoded in the parent `linkerd install` output, via
[Cluster]Role[Binding]s.

Modify the control-plane components to query Kubernetes at startup to
determine which namespaces they are authorized to access, and whether
ServiceProfile support is available. This allows removal of the
`--single-namespace` flag on the components.

Also update `bin/test-cleanup` to cleanup the ServiceProfile CRD.

TODO:
- Remove `--single-namespace` flag on `linkerd install`, part of #2164

Signed-off-by: Andrew Seigner <siggy@buoyant.io>